feat: add duplicate check validator with test case

+++ pkg/hook.go
@@ -5,11 +5,8 @@ import (
 	"fmt"
 	"log"
 	"os"
-
-	// "path"
 	"strings"
 
-	"github.com/google/uuid"
 	constant "github.com/hexbook/internal/constant"
 	qrcode "github.com/skip2/go-qrcode"
 )
@@ -18,36 +15,66 @@ func validateAddress(address string) {
 	_, found := strings.CutPrefix(address, "0x")
 
 	if !found || len(address) != 42 {
-		error := errors.New("validateAddress.go: invalid ethereum address\n")
-		log.Fatalln(error.Error(), "address length: ", len(address))
+		error := errors.New("validateAddress.go: invalid ethereum address")
+		log.Fatalln(error.Error(), "| address length: ", len(address))
 	}
+}
+
+func validateDuplicate(address string) bool {
+	wd, gErr := os.Getwd()
+
+	if gErr != nil {
+		log.Fatalln("validateDuplicate: ", gErr.Error())
+	}
+
+	targetPath := strings.Join([]string{wd, "assets", "qrcode"}, "/")
+	entries, rErr := os.ReadDir(targetPath)
+
+	if rErr != nil {
+		log.Fatalln("validateDuplicate: ", rErr.Error())
+	}
+
+	var isExisting bool = false
+
+	for _, v := range entries {
+		if v.Name() == address {
+			isExisting = true
+			break
+		}
+	}
+
+	return isExisting
 }
 
 func GenerateQrCode(wallet string) string {
 	validateAddress(wallet)
-	png, err := qrcode.Encode(wallet, qrcode.Medium, 256)
+	isExisting := validateDuplicate(wallet)
+	filename := fmt.Sprintf("%s.png", wallet)
 
-	if err != nil {
-		log.Fatalln(err.Error())
+	if !isExisting {
+		log.Println("GenerateQrCode: detecting new entry for qrcode, starting encoding...", filename)
+		png, err := qrcode.Encode(wallet, qrcode.Medium, 256)
+
+		if err != nil {
+			log.Fatalln(err.Error())
+		}
+
+		wd, gErr := os.Getwd()
+
+		if gErr != nil {
+			log.Fatalln(gErr.Error())
+		}
+
+		targetPath := strings.Join([]string{wd, "assets", "qrcode", filename}, "/")
+		wErr := os.WriteFile(targetPath, png, constant.FilePermUserReadWriteGroupRead)
+
+		if wErr != nil {
+			log.Fatalln(wErr.Error())
+		}
+
+		return filename
 	}
 
-	wd, gErr := os.Getwd()
-
-	if gErr != nil {
-		log.Fatalln(gErr.Error())
-	}
-
-	id := uuid.New().String()
-	filename := fmt.Sprintf("%s.png", id)
-
-	log.Println("wd: ", wd, "filename: ", filename)
-
-	targetPath := strings.Join([]string{wd, "assets", "qrcode", filename}, "/")
-	wErr := os.WriteFile(targetPath, png, constant.FilePermUserReadWriteGroupRead)
-
-	if wErr != nil {
-		log.Fatalln(wErr.Error())
-	}
-
+	log.Println("GenerateQrCode: detecting existing entry for qrcode, terminating...", filename)
 	return filename
 }
